% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocktailApp.r
\docType{data}
\name{cocktails}
\alias{cocktails}
\title{Cocktails Data}
\format{A \code{data.frame} object with 77,301 rows and 12 columns. The
data are scraped from three websites: Difford's guide, Webtender, and 
Kindred Cocktails.

The columns are defined as follows:
\describe{
 \item{\code{amt}}{The numeric amount of the ingredient.}
 \item{\code{unit}}{The unit corresponding to the amount. Most have been converted to \code{fl oz}.}
 \item{\code{ingredient}}{The name of the ingredient.}
 \item{\code{cocktail}}{The name of the cocktail.}
 \item{\code{rating}}{The rating assigned to the cocktail in the upstream database.}
 \item{\code{upstream_id}}{An ID code from the upstream source.}
 \item{\code{url}}{The upstream URL.}
 \item{\code{votes}}{The number of votes in the rating. Not always available.}
 \item{\code{added}}{The date the cocktail was added to the upstream database. Not always available.}
 \item{\code{short_ingredient}}{A shortened form of the ingredient. This is subject to change, as better term extraction is needed.}
 \item{\code{proportion}}{For non-garnish ingredients, this is the proportion of the given cocktail that consists of the given ingredient. This
 is a normalized amount.}
}}
\source{
Difford's Guide, \url{http://www.diffordsguide.com/},
Webtender, \url{http://www.webtender.com},
Kindred Cocktails, \url{http://kindredcocktails.com}.
}
\usage{
cocktails
}
\description{
Ingredients of several thousand cocktails, scraped from the web.
}
\examples{
\dontrun{
data(cocktails)
str(cocktails)
}
}
\author{
Steven E. Pav \email{steven@gilgamath.com}
}
\keyword{datasets}
