% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocktailApp.r
\docType{package}
\name{cocktailApp-package}
\alias{cocktailApp-package}
\title{Shiny app to discover cocktails.}
\description{
Shiny app to discover cocktails.
}
\note{
This package is maintained as a hobby.
}
\section{Legal Mumbo Jumbo}{


cocktailApp is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU Lesser General Public License for more details.
}

\author{
Steven E. Pav \email{shabbychef@gmail.com}
}
\keyword{package}
