% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocktailApp.r
\docType{package}
\name{cocktailApp}
\alias{cocktailApp}
\alias{cocktailApp-package}
\alias{cocktailApp}
\title{Shiny app to discover cocktails.}
\usage{
cocktailApp()
}
\value{
Runs the shiny app.
}
\description{
Shiny app to discover cocktails.

One sentence or so that tells you some more.
}
\note{
This package is maintained as a hobby.
}
\section{Legal Mumbo Jumbo}{


cocktailApp is distributed in the hope that it will be useful,
but WITHOUT ANY WARRANTY; without even the implied warranty of
MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
GNU Lesser General Public License for more details.
}

\examples{
\dontrun{
cocktailApp()
}
}
\author{
Steven E. Pav \email{shabbychef@gmail.com}

Steven E. Pav \email{shabbychef@gmail.com}

Steven E. Pav \email{steven@gilgamath.com}
}
\keyword{package}
\keyword{shiny}
