% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocktailApp.r
\name{cocktailApp-NEWS}
\alias{cocktailApp-NEWS}
\title{News for package 'cocktailApp':}
\description{
News for package \sQuote{cocktailApp}

\newcommand{\CRANpkg}{\href{https://cran.r-project.org/package=#1}{\pkg{#1}}}
\newcommand{\cocktailApp}{\CRANpkg{cocktailApp}}
}
\section{\cocktailApp{} Initial Version 0.1.0 (2018-07-05) }{

\itemize{
\item first CRAN release.
}
}

