% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocktailApp.r
\name{cocktailApp-NEWS}
\alias{cocktailApp-NEWS}
\title{News for package 'cocktailApp':}
\description{
News for package \sQuote{cocktailApp}

\newcommand{\CRANpkg}{\href{https://cran.r-project.org/package=#1}{\pkg{#1}}}
\newcommand{\cocktailApp}{\CRANpkg{cocktailApp}}
}
\section{\cocktailApp{} Version 0.2.4 (2025-11-04) }{

\itemize{
\item removing dependency on ggtern.
\item adding missing reference to `dplyr::n` which broke two of the plots.
}
}

\section{\cocktailApp{} Version 0.2.3 (2023-07-18) }{

\itemize{
\item fix bug where you could not search for Gin or Vodka.
\item package had been archived because of dependency on ggtern.
\item convert citEntry to bibentry
}
}

\section{\cocktailApp{} Version 0.2.2 (2021-04-01) }{

\itemize{
\item merge many short ingredients.
\item fix some units in webtender data.
\item search by ingredient regex.
}
}

\section{\cocktailApp{} Version 0.2.1 (2019-07-01) }{

\itemize{
\item CRAN fix as tests were hanging.
\item replace \code{Ternary} package with \code{ggtern}.
}
}

\section{\cocktailApp{} Version 0.2.0 (2018-08-19) }{

\itemize{
\item adding another source.
\item adding \dQuote{Hobson's Choice} button.
\item removing dependency on \code{ggtern} package, replacing with \code{Ternary}.
}
}

\section{\cocktailApp{} Initial Version 0.1.0 (2018-07-05) }{

\itemize{
\item first CRAN release.
}
}

